Centralize folder/domain field rendering in ModelForm.svelte
Moves the FolderTreeSelect (domain picker) and the folder/perimeter cascade logic from ~50 individual form components into ModelForm.svelte, alongside the already-centralized ref_id/name/description triplet.

What changed:

ModelForm.svelte: renders FolderTreeSelect when shape.folder exists, plus the perimeter autocomplete when shape.perimeter exists. Includes cascade logic (clear perimeter on folder change, auto-fill folder from perimeter selection). Adds contentTypes={['DO', 'GL', 'EN']} for compliance-assessments and evidences to include enclave folders.
5 complex forms (RiskAssessment, ComplianceAssessment, BusinessImpactAnalysis, EntityAssessment, FindingsAssessment): removed ~65 lines each of duplicated folder/perimeter cascade logic.
~45 simple forms: removed FolderTreeSelect import + template block (~8 lines each).
ValidationFlowForm: keeps its own FolderTreeSelect (custom fetchDefaultRefId on folder change).
EvidenceRevisionForm: folder rendered as HiddenInput (inherited from parent evidence, not user-selectable).
Field order: ref_id+name / description / folder / perimeter / model-specific fields.
