+++ .pnpm-store/v11/.tmp/pnpm-10.33.2-1781998042774/package.json
@@ -1,0 +1,1 @@
+{"dependencies":{"pnpm":"10.33.2"}}
