+++ frontend/src/lib/components/Forms/ModelForm/AttackPathForm.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import AutocompleteSelect from '$lib/components/Forms/AutocompleteSelect.svelte';
-	import FolderTreeSelect from '$lib/components/Forms/FolderTreeSelect.svelte';
 	import type { CacheLock, ModelInfo } from '$lib/utils/types';
 	import { m } from '$paraglide/messages';
 	import type { SuperForm } from 'sveltekit-superforms';
@@ -38,14 +37,6 @@
 	bind:cachedValue={formDataCache['strategic_scenario']}
 	label={m.strategicScenario()}
 	hidden={initialData['strategic_scenario']}
-/>
-<FolderTreeSelect
-	{form}
-	field="folder"
-	cacheLock={cacheLocks['folder']}
-	bind:cachedValue={formDataCache['folder']}
-	label={m.folder()}
-	hidden
 />
 <AutocompleteSelect
 	{form}
