+++ frontend/src/lib/components/Forms/ModelForm/EntityForm.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import AutocompleteSelect from '../AutocompleteSelect.svelte';
-	import FolderTreeSelect from '../FolderTreeSelect.svelte';
 	import TextArea from '$lib/components/Forms/TextArea.svelte';
 	import TextField from '$lib/components/Forms/TextField.svelte';
 	import NumberField from '$lib/components/Forms/NumberField.svelte';
@@ -58,14 +57,6 @@
 	label={m.isActive()}
 	cacheLock={cacheLocks['is_active']}
 	bind:cachedValue={formDataCache['is_active']}
-/>
-<FolderTreeSelect
-	{form}
-	field="folder"
-	cacheLock={cacheLocks['folder']}
-	bind:cachedValue={formDataCache['folder']}
-	label={m.domain()}
-	disabled={object.builtin}
 />
 {#if !object.builtin}
 	<AutocompleteSelect
