+++ frontend/src/lib/components/Forms/ModelForm/EvidenceForm.svelte
@@ -2,7 +2,6 @@
 	import HiddenInput from '../HiddenInput.svelte';
 	import FileInput from '../FileInput.svelte';
 	import AutocompleteSelect from '../AutocompleteSelect.svelte';
-	import FolderTreeSelect from '../FolderTreeSelect.svelte';
 	import TextField from '$lib/components/Forms/TextField.svelte';
 	import Select from '$lib/components/Forms/Select.svelte';
 	import type { SuperValidated } from 'sveltekit-superforms';
@@ -72,19 +71,6 @@
 		allowedExtensions={'*'}
 	/>
 {/if}
-<FolderTreeSelect
-	{form}
-	field="folder"
-	cacheLock={cacheLocks['folder']}
-	bind:cachedValue={formDataCache['folder']}
-	label={m.domain()}
-	optionsInfoFields={{
-		fields: [{ field: 'content_type', translate: true }],
-		position: 'suffix',
-		classes: 'text-xxs bg-surface-200-800 px-1.5 py-0.5 rounded'
-	}}
-	contentTypes={['DO', 'GL', 'EN']}
-/>
 {#if context !== 'edit'}
 	<TextField
 		{form}
