+++ frontend/src/lib/components/Forms/ModelForm/MetricDefinitionForm.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import AutocompleteSelect from '../AutocompleteSelect.svelte';
-	import FolderTreeSelect from '../FolderTreeSelect.svelte';
 	import TextField from '$lib/components/Forms/TextField.svelte';
 	import TextArea from '$lib/components/Forms/TextArea.svelte';
 	import Checkbox from '$lib/components/Forms/Checkbox.svelte';
@@ -61,13 +60,6 @@
 	}
 </script>
 
-<FolderTreeSelect
-	{form}
-	field="folder"
-	cacheLock={cacheLocks['folder']}
-	bind:cachedValue={formDataCache['folder']}
-	label={m.domain()}
-/>
 <Select
 	{form}
 	options={model.selectOptions['category']}
