+++ frontend/src/lib/components/Forms/ModelForm/OrganisationIssueForm.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import AutocompleteSelect from '../AutocompleteSelect.svelte';
-	import FolderTreeSelect from '../FolderTreeSelect.svelte';
 	import TextField from '$lib/components/Forms/TextField.svelte';
 	import TextArea from '$lib/components/Forms/TextArea.svelte';
 	import Select from '../Select.svelte';
@@ -78,13 +77,6 @@
 	helpText={m.observationHelpText()}
 	cacheLock={cacheLocks['observation']}
 	bind:cachedValue={formDataCache['observation']}
-/>
-<FolderTreeSelect
-	{form}
-	field="folder"
-	cacheLock={cacheLocks['folder']}
-	bind:cachedValue={formDataCache['folder']}
-	label={m.domain()}
 />
 <TextField
 	type="date"
