+++ frontend/src/lib/components/Forms/ModelForm/ProjectForm.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import AutocompleteSelect from '../AutocompleteSelect.svelte';
-	import FolderTreeSelect from '../FolderTreeSelect.svelte';
 	import Select from '../Select.svelte';
 	import CustomFieldsSection from '../CustomFieldsSection.svelte';
 	import type { CacheLock, ModelInfo } from '$lib/utils/types';
@@ -37,14 +36,6 @@
 	cacheLock={cacheLocks['kind']}
 	bind:cachedValue={formDataCache['kind']}
 	disableDoubleDash={true}
-/>
-
-<FolderTreeSelect
-	{form}
-	field="folder"
-	cacheLock={cacheLocks['folder']}
-	bind:cachedValue={formDataCache['folder']}
-	label={m.domain()}
 />
 
 <AutocompleteSelect
