+++ frontend/src/lib/components/Forms/ModelForm/QuantitativeRiskStudyForm.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import AutocompleteSelect from '../AutocompleteSelect.svelte';
-	import FolderTreeSelect from '../FolderTreeSelect.svelte';
 	import TextField from '$lib/components/Forms/TextField.svelte';
 	import NumberField from '$lib/components/Forms/NumberField.svelte';
 	import MarkdownField from '$lib/components/Forms/MarkdownField.svelte';
@@ -106,14 +105,6 @@
 		}
 	}
 </script>
-
-<FolderTreeSelect
-	{form}
-	field="folder"
-	cacheLock={cacheLocks['folder']}
-	bind:cachedValue={formDataCache['folder']}
-	label={m.domain()}
-/>
 
 <AutocompleteSelect
 	{form}
